requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
typing
hydra-core
numpy<=1.26.4
torch==2.0.1
igraph
scipy
transformers==4.40.1
evaluate
pillow
pandas
joblib